requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
